// src/store/ui/interfaces/ui-state.interface.ts
export interface IState {
  navbarDropdownOpen: boolean;
  sidebarSelectedItem: string;
  excerciseSelectedItem: number | null;
  menuExcerciseSelectedItem: number | null;
  menuEvaluationSelectedItem: number | null;
  menuClassroomSelectedItem: number | null;
  messagesSelectedItem: string;
  clock: string;
  selectedClassroom: string | null;
}